Separate suites into sep files in /suites.
