+++ tests/10-bbs-create.js
@@ -3,7 +3,7 @@
  *
  * SPDX-License-Identifier: BSD-3-Clause
  */
-import {createSuite} from './suites.js';
+import {createSuite} from './suites/create.js';
 import {endpoints} from 'vc-test-suite-implementations';
 import {getSuiteConfig} from './test-config.js';
